support title, resolve sockethangup issues, resolve process control issue and enhance alert element search
