+++ XCTestWD/XCTestWDUITests/server/controllers/XCTestWDElementController.swift
@@ -30,6 +30,7 @@ internal class XCTestWDElementController: Controller {
                 (RequestRoute("/wd/hub/session/:sessionId/tap/:elementId", "post"), tap),
                 (RequestRoute("/wd/hub/session/:sessionId/doubleTap", "post"), doubleTapAtCoordinate),
                 (RequestRoute("/wd/hub/session/:sessionId/keys", "post"), handleKeys),
+                (RequestRoute("/wd/hub/session/:sessionId/title", "get"), title),
                 (RequestRoute("/wd/hub/session/:sessionId/homeScreen", "post"), homeScreen),
                 (RequestRoute("/wd/hub/session/:sessionId/element/:elementId/doubleTap", "post"), doubleTap),
                 (RequestRoute("/wd/hub/session/:sessionId/element/:elementId/touchAndHold", "post"), touchAndHoldOnElement),
@@ -428,7 +429,29 @@ internal class XCTestWDElementController: Controller {
         return XCTestWDResponse.response(session: nil, error: WDStatus.Success)
     }
     
+    internal static func title(request: Swifter.HttpRequest) -> Swifter.HttpResponse {
+        
+        let session = XCTestWDSessionManager.singleton.checkDefaultSession()
+        let application = session.application
+        let elements = application?.descendants(matching: XCUIElementType.window).allElementsBoundByIndex
+        
+        if  elements == nil || elements?.count == 0 {
+            return XCTestWDResponse.response(session: nil, error: WDStatus.ElementNotVisible)
+        }
+        
+        let window = elements![0]
+        let navBar = window.descendants(matching: XCUIElementType.navigationBar).allElementsBoundByIndex.first
+        
+        if (navBar?.identifier.characters.count) ?? 0 > 0 {
+            return XCTestWDResponse.response(session: nil, value: JSON(navBar?.identifier as Any))
+        } else {
+            window.resolve()
+            let digest = window.digest()
+            return XCTestWDResponse.response(session: nil, value: JSON(digest as Any))
+        }
+    }
     
+
     internal static func doubleTap(request: Swifter.HttpRequest) -> Swifter.HttpResponse {
         let elementId = request.elementId
         let session = request.session ?? XCTestWDSessionManager.singleton.checkDefaultSession()
