+++ XCTestWD/XCTestWDUITests/server/models/utils/XCTestWDMathUtils.swift
@@ -23,5 +23,18 @@ class MathUtils {
         }
         return actualSize;
     }
+    
+    static func MD5(string: String) -> String {
+        let messageData = string.data(using:.utf8)!
+        var digestData = Data(count: Int(CC_MD5_DIGEST_LENGTH))
+        
+        _ = digestData.withUnsafeMutableBytes {digestBytes in
+            messageData.withUnsafeBytes {messageBytes in
+                CC_MD5(messageBytes, CC_LONG(messageData.count), digestBytes)
+            }
+        }
+        
+        return digestData.base64EncodedString()
+    }
 
 }
